Fixed Swift Sauce.
Updated movement script so it updates which stat it uses for the movement.
Swift sauce now when used, will not be droppable for the duration of the item effect; this means that it will lock the item slot out until duration is expended. When a player switch speed will no longer be transfered to the swapped player.